+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 - Support disk usage reporting (using `df`) on Alpine Linux.
-- Don't report inodes for mountpoints that report no usage percentage.
+- When a disk mountpoint has no inodes usage percentage, skip the mountpoint, and report the inodes information successfully for the inodes that do have an inodes usage percentage.
 
 ## 0.5.2